Move nodecustomdata renderer tests into baker_test.go
Keep renderer coverage alongside baker.go after the renderer implementation was consolidated there.


Copilot-Session: 4ae1dd11-ec06-4c26-bfc8-d1ef1f89fcf9
